# src/main/resources/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: API de Geolocalização
  version: 1.0.0
  description: Sistema de Geolocalização.

servers:
  - url: http://localhost:8080

paths:
  /v1/findLocation:
    get:
      summary: Pesquisa de logradouros por CEP.
      operationId: searchPlaceByCep
      tags:
        - Search
      parameters:
        - in: query
          name: cep
          schema:
            type: string
            pattern: '^[0-9]{8}$'
          required: true
      responses:
        '200':
          description: Local encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPlaceResponse'
        '400':
          description: Parâmetro 'cep' inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '404':
          description: CEP não encontrado na API externa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '503':
          description: Erro de comunicação com o serviço externo de CEP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro inesperado no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'

  /v1/location:
    get:
      summary: Pesquisa de locais cadastrados por nome ou geolocalização.
      operationId: searchPlaceByQuery
      tags:
        - Search
      parameters:
        - name: nome
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 50
        - name: latitude
          in: query
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          schema:
            type: number
            format: double
        - name: raio
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Busca realizada com sucesso (pode retornar uma lista vazia).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlaceResponse'
        '400':
          description: Parâmetros de busca inválidos ou inconsistentes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro inesperado no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'

    post:
      summary: Cadastrar um novo local.
      operationId: insertNewPlace
      tags:
        - Insert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPlaceRequest'
      responses:
        '201':
          description: Local registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceResponse'
        '400':
          description: Payload da requisição inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '404':
          description: CEP informado no payload não foi encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '409':
          description: Já existe um local cadastrado com os dados informados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '503':
          description: Erro de comunicação com o serviço externo de CEP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro inesperado no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'

components:
  schemas:
    PlaceResponse:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        complemento:
          type: string
          nullable: true
        cep:
          type: string
        rua:
          type: string
        numero:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        geolocalizacao:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
        avaliacao:
          type: number
          format: double

    SearchPlaceResponse:
      type: object
      properties:
        cep:
          type: string
        rua:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        geolocalizacao:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double

    NewPlaceRequest:
      type: object
      properties:
        cep:
          type: string
          pattern: '^[0-9]{8}$'
        nome:
          type: string
          minLength : 2
          maxLength: 50
        complemento:
          type: string
          nullable: true
        numero:
          type: string
          minLength: 1
          maxLength: 7
        avaliacao:
          type: number
          format: double
          minimum: 0
          maximum: 5
      required:
        - cep
        - nome
        - numero
        - avaliacao

    StandardError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
      required:
        - timestamp
        - status
        - error